+++ packages/config/src/index.ts
@@ -1,0 +1,6 @@
+export * from './Modules';
+export * from './Web';
+export * from './Config';
+export * from './Config.types';
+export * from './Project';
+export * from './Errors';
